refactor(tool): group tools by domain not verb
- read.py/write.py split by http verb forced unrelated tools together (delete_work_item_links sat in "write")
- replace with projects/work_items/documents/links/comments/moves modules over a shared _shared/ layer
