+++ src/mcp_server_polarion/tools/__init__.py
@@ -1,8 +1,15 @@
-"""MCP tool definitions — read and write tools for Polarion ALM."""
+"""MCP tool definitions — domain-grouped tools for Polarion ALM.
+
+Importing each module registers its ``@mcp.tool`` functions as a side effect.
+"""
 
 from __future__ import annotations
 
-import mcp_server_polarion.tools.read
-import mcp_server_polarion.tools.write  # noqa: F401
+import mcp_server_polarion.tools.comments
+import mcp_server_polarion.tools.documents
+import mcp_server_polarion.tools.links
+import mcp_server_polarion.tools.moves
+import mcp_server_polarion.tools.projects
+import mcp_server_polarion.tools.work_items  # noqa: F401
 
 __all__: list[str] = []
