+++ src/mcp_server_polarion/tools/_shared/__init__.py
@@ -1,0 +1,1 @@
+"""Shared helpers, cache, and write guards used across tool modules."""
